// SDKs/twilio.com/twilio_flex_v1/1.37.3/typescript/src/sdk/models/shared/flexv1webchannel.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class FlexV1WebChannel extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=flex_flow_sid" })
  flexFlowSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}